+++ cucumber/cucumber-runner/src/executor/test-executor.ts
@@ -21,6 +21,7 @@ import { globalScope } from "@scopes/globals";
 export class TestExecutor {
   subscribers: ProviderSubscriber[];
   #instanceDependencies: DependencyInstanceProvider[];
+  #globalApp: unknown;
   constructor(private feature: GherkinFeature) {
     const prototypes = Config.get<Class<ProviderSubscriber>[]>("subscribers");
     this.subscribers = prototypes?.map((it) => new it()) ?? [];
@@ -31,6 +32,7 @@ export class TestExecutor {
       ?.flatMap<DependencyInstanceProvider>(
         (sub) => sub.fixtures.instances as DependencyInstanceProvider[]
       );
+    this.#globalApp = getApp(...this.#instanceDependencies);
   }
   execute() {
     const { beforeAll, afterAll, describe } = Config.get<TestFunctions>("runner");
@@ -44,11 +46,11 @@ export class TestExecutor {
     beforeAll(async () => {
       events.feature.emitStart({ title, path, tags, modifier });
       const setup = [...globalScope.hooks?.setup, ...this.feature.hooks?.setup];
-      await runSetupHooks(setup, events.setup, failFeature);
+      await runSetupHooks(setup, this.#globalApp, failFeature);
     });
     afterAll(async () => {
       const teardown = [...globalScope.hooks?.teardown, ...this.feature.hooks?.teardown];
-      await runTeardownHooks(teardown, failFeature);
+      await runTeardownHooks(teardown, this.#globalApp, failFeature);
     });
     afterAll(() => {
       events.feature.emitEnd({ title, status: failed ? Status.FAILED : Status.PASSED });
@@ -89,11 +91,11 @@ export class TestExecutor {
       }
       beforeAll(async () => {
         events.rule.emitStart({ title, modifier, tags });
-        await runSetupHooks(rule.hooks?.setup, events.setup, failRule);
+        await runSetupHooks(rule.hooks?.setup, this.#globalApp, failRule);
       });
 
       afterAll(async () => {
-        await runTeardownHooks(this.feature.hooks?.teardown, onFailure);
+        await runTeardownHooks(this.feature.hooks?.teardown, this.#globalApp, onFailure);
 
         events.rule.emitEnd({ title, status: failed ? Status.FAILED : Status.PASSED });
       });
@@ -123,11 +125,11 @@ export class TestExecutor {
       }
       beforeAll(async () => {
         events.scenarioOutline.emitStart({ title, tags, modifier, examples });
-        await runSetupHooks(hooks?.setup, events.setup, failOutline);
+        await runSetupHooks(hooks?.setup, this.#globalApp, failOutline);
       });
 
       afterAll(async () => {
-        await runTeardownHooks(outline.hooks?.teardown, onFailure);
+        await runTeardownHooks(outline.hooks?.teardown, this.#globalApp, onFailure);
         events.scenarioOutline.emitEnd({ title, status: failed ? Status.FAILED : Status.PASSED });
       });
       const scenarios = outline.scenarios;
@@ -150,7 +152,7 @@ export class TestExecutor {
     const testFn = tagFilter(scenario.tags, test as FrameworkTestCall, scenario.modifier);
     testFn(scenario.getScenarioTitle(), async () => {
       events.scenarioWrapper.emitStart();
-      await runBeforeHooks(befores, app, onFailure);
+      await runBeforeHooks(befores, scenario.tags, app, onFailure);
       await runBackgrounds(scenario, app);
       try {
         events.scenario.emitStart({
@@ -169,7 +171,7 @@ export class TestExecutor {
         events.scenario.emitEnd({ title, status: Status.FAILED, modifier, error: e as Error });
         throw e;
       } finally {
-        await runAfterHooks(afters, app, onFailure);
+        await runAfterHooks(afters, scenario.tags, app, onFailure);
         events.scenarioWrapper.emitEnd();
       }
     });
@@ -228,8 +230,16 @@ async function runSteps(
   }
 }
 
-async function runBeforeHooks(befores: BeforeHook[], app: unknown, onFailure: OnFailure) {
+async function runBeforeHooks(
+  befores: BeforeHook[],
+  tags: string[],
+  app: unknown,
+  onFailure: OnFailure
+) {
   for (const hook of befores ?? []) {
+    if (!isTagFiltered(tags, hook.tagFilter)) {
+      continue;
+    }
     const { description } = hook;
     events.before.emitStart({ description });
     try {
@@ -261,8 +271,16 @@ async function runSetupHooks(setups: SetupHook[], app: unknown, onFailure: OnFai
   }
 }
 
-async function runAfterHooks(afters: AfterHook[], app: unknown, onFailure: OnFailure) {
+async function runAfterHooks(
+  afters: AfterHook[],
+  tags: string[],
+  app: unknown,
+  onFailure: OnFailure
+) {
   for (const hook of afters?.reverse() ?? []) {
+    if (!isTagFiltered(tags, hook.tagFilter)) {
+      continue;
+    }
     const { description } = hook;
     events.after.emitStart({ description });
     try {
@@ -278,12 +296,12 @@ async function runAfterHooks(afters: AfterHook[], app: unknown, onFailure: OnFai
   }
 }
 
-async function runTeardownHooks(teardowns: TeardownHook[], onFailure: OnFailure) {
+async function runTeardownHooks(teardowns: TeardownHook[], app: unknown, onFailure: OnFailure) {
   for (const hook of teardowns?.reverse() ?? []) {
     const { description } = hook;
     events.teardown.emitStart({ description });
     try {
-      await hook.action();
+      await hook.action(app);
       events.teardown.emitEnd({ description, status: Status.PASSED });
     } catch (e) {
       const oldMessage = (e as Error).message;
@@ -346,4 +364,14 @@ function tagFilter(tags: string[], fn: FrameworkTestCall | TestGroup, modifiers?
   }
 
   return fn;
+}
+
+export function isTagFiltered(tags: string[], filter?: string) {
+  if (filter) {
+    if (!parse(filter).evaluate(tags)) {
+      return false;
+    }
+  }
+
+  return true;
 }