+++ cucumber/integration-tests-esm/test/hooks.feature.ts
@@ -36,7 +36,7 @@ const results = {
     },
   },
 };
-Setup("Outer setup", () => {
+Setup("Outer setup", (app) => {
   results.base.setupCalled = true;
 });
 Before("Outer before", () => {
@@ -45,11 +45,12 @@ Before("Outer before", () => {
 After("Outer After", () => {
   results.base.afterCalled = true;
 });
-Teardown("Outer Teardown", () => {
+Teardown("Outer Teardown", (app) => {
   results.base.tearDownCalled = true;
   expect(results.OuterOutline.tearDownCalled).toBe(true);
   expect(results.rule.ruleOutline.tearDownCalled).toBe(true);
 });
+After("beardown", () => console.error("Running"), "@skipHook");
 Feature(() => {
   Scenario("Outer Hook Scenario", () => {
     Given("a scenario with hooks", () => {