+++ packages/events/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Gherkin
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [64bacd4]
+  - @autometa/gherkin@0.2.3
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [064b589]
+  - @autometa/dto-builder@0.8.0
+  - @autometa/gherkin@0.2.2
+
 ## 0.0.2
 
 ### Patch Changes